color switch thing that maam asked for
paayos na lang ng colors if di makita, under colors.xml, mood_x_darker

+++ app/src/main/java/com/mobdeve/s14/group24/everyday/EditMediaEntryActivity.java
@@ -156,7 +156,8 @@ public class EditMediaEntryActivity extends AppCompatActivity {
                     @Override
                     public void run() {
                         databaseHelper.updateData(id, imagePath, caption, mood);
-                        new File(oldImagePath).delete();
+                        if (imagePath != oldImagePath)
+                            new File(oldImagePath).delete();
                     }
                 });
